requirements.txt
# Web API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

# Database
sqlalchemy>=2.0.35
psycopg2-binary>=2.9.10

# Auth & Security
python-jose[cryptography]>=3.3.0
bcrypt>=4.0.0

# Validation & Settings
pydantic>=2.9.2
pydantic-settings>=2.5.2

# Environment
python-dotenv>=1.0.1

# Data & ML (sin pines estrictos para compatibilidad con Python 3.14)
pandas>=2.2.3
numpy>=2.0.0
lightgbm>=4.5.0
scikit-learn>=1.5.2
mlxtend>=0.23.1

# Prophet requiere Python <= 3.12
# prophet==1.3.0
